Moved method detection helpers to base/eos_concepts.hpp
